google-adk/firestore_test_connect/requirements.txt
google-cloud-storage>=2.10.0
python-dotenv>=1.0.0
firebase-admin>=6.5.0
requests>=2.31.0
google-generativeai>=0.8.0
pydantic>=2.0.0
cloudpickle>=3.0.0
deprecated>=1.2.0
google-cloud-aiplatform[adk,agent_engines]>=1.50.0
